refactor cloud watch logs put log event action
